edit icon and stuffs
