+++ app/client/src/widgets/TableWidgetV2/component/cellComponents/TextCell.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback } from "react";
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 import AutoToolTipComponent from "../AutoToolTipComponent";
 import { renderDefaultPropsType } from "../renderHelpers/DefaultRenderer";
 import { ReactComponent as EditIcon } from "assets/icons/control/edit-variant1.svg";
@@ -11,15 +11,29 @@ const Wrapper = styled.div`
   width: 100%;
 `;
 
-const EDIT_ICON_WIDTH = "20px";
+const EDIT_ICON_WIDTH = "30px";
 
 const StyledAutoToolTipComponent = styled(AutoToolTipComponent)`
-  flex-basis: calc(100% - ${EDIT_ICON_WIDTH});
+  width: 100%;
 `;
 
 const StyledEditIcon = styled(EditIcon)`
   flex-basis: ${EDIT_ICON_WIDTH};
   cursor: pointer;
+  height: 10px;
+  display: none;
+`;
+
+export const EditableRowHoverStyle = css`
+  &:hover {
+    .editable-cell {
+      width: calc(100% - ${EDIT_ICON_WIDTH});
+    }
+
+    .editable-cell-icon {
+      display: block;
+    }
+  }
 `;
 
 type TextCellProps = Omit<
@@ -30,6 +44,7 @@ type TextCellProps = Omit<
 export function TextCell({
   cellProperties,
   columnType,
+  isCellEditable,
   isCellVisible,
   isHidden,
   tableWidth,
@@ -48,15 +63,18 @@ export function TextCell({
     <Wrapper>
       <StyledAutoToolTipComponent
         cellProperties={cellProperties}
+        className={isCellEditable ? "editable-cell" : ""}
         columnType={columnType}
         isCellVisible={isCellVisible}
         isHidden={isHidden}
         tableWidth={tableWidth}
         title={!!value ? value.toString() : ""}
       >
-        <div onClick={() => toggleCellEditMode(true)}>{value}</div>
+        {value}
       </StyledAutoToolTipComponent>
-      <StyledEditIcon onClick={onIconClick} />
+      {isCellEditable && (
+        <StyledEditIcon className="editable-cell-icon" onClick={onIconClick} />
+      )}
     </Wrapper>
   );
 }
