+++ app/client/src/widgets/TableWidgetV2/widget/index.tsx
@@ -46,6 +46,7 @@ import {
   getSelectRowIndex,
   getSelectRowIndices,
   getHash,
+  getCellProperties,
 } from "./utilities";
 
 import {
@@ -58,7 +59,6 @@ import {
 import tablePropertyPaneConfig from "./propertyConfig";
 import { BatchPropertyUpdatePayload } from "actions/controlActions";
 import { IconName } from "@blueprintjs/icons";
-import { getCellProperties } from "./getTableColumns";
 import { Colors } from "constants/Colors";
 import { IconNames } from "@blueprintjs/core/node_modules/@blueprintjs/icons";
 import equal from "fast-deep-equal/es6";
